Update strings.xml
Few corrections and additions for the Italian translation that sound more natural. Also, before merging, I would like to understand what's the context for the "show_grouped_calls" string. I couldn't find anything in the app, and naturally the translation changes based on the context. Furthermore, the translation for the speed dial section is completely wrong, so I took the liberty to change a few strings.